pyproject.toml
[project]
name = "finbound"
version = "0.1.0"
description = "Minimal FinBound prototype: verification-gated financial reasoning pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "FinBound Team" }
]

dependencies = [
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["finbound*"]
